requirements.txt
alembic==1.10.2
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pyngrok==5.2.1
pytest==7.2.2
python-dotenv==1.0.0
pytz==2023.2
PyYAML==6.0
requests==2.28.2
SQLAlchemy==2.0.4
stripe==5.2.0
tomli==2.0.1
twilio==7.17.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
